refactor: add CommandsMixin and modularize sensor, property, and manager logic
